refactor: Move AuthProvider to context directory and update imports
- Relocate AuthProvider to a new context directory for better organization
- Update import paths in page and ProviderWrapper components to reflect the new structure
- Introduce AuthProvider implementation for managing user authentication state
